move timeserieslong transform to the right spot
